Updated PuzzleGUI
Figured I should submit this as a request since it is a large change.

I simplified the showXXXX() methods to just remove then add the new panel.
Organized general variable and component placement.
Customized the buttons to have their own size and font.
Set the layout so that the intro screen buttons have their proper placement.
Created areas for future panel sections.

NOTE: Lines 111 - 113 are commented out since they won't work for everyone because of the image location. For an example of how it would make the window look you can check out the Skype chat or the Week 5 folder in Google Drive.

Let me know if this is too much or too little. We can also set the comments to how they were before. I didn't mind them. I just removed them to make it simpler to read while I was coding.